IS1MA-71 Reestructura archivos y modificaciones en crud Player
Reestructuracion de archivos acorde a la rama develop.

Modificaciones en las funciones crud de Player para un mejor manejo de excepciones y correcta
devolucion de valores de retorno

+++ src/theThing/players/crud.py
@@ -1,9 +1,12 @@
-from src.games import crud as games_crud
+from pony.orm import ObjectNotFound
 from pony.orm import db_session
-from src.players.schemas import PlayerCreate, PlayerUpdate, PlayerBase
+
+from src.theThing.games.models import Game
+from src.theThing.players.schemas import PlayerCreate, PlayerUpdate, PlayerBase
+from .models import Player
 
 
-def create_player(player: PlayerCreate, game: int):
+def create_player(player_data: PlayerCreate, game_id: int):
     """
     It creates a player in the database from the
     PlayerCreate schema and returns the PlayerBase schema
@@ -13,20 +16,20 @@ def create_player(player: PlayerCreate, game: int):
     Then an exception its raised.
     """
     with db_session:
-        game_to_join = games_crud.get_game(game)
+        game_to_join = Game[game_id]
         if game_to_join.state != 0:
             raise Exception("Game already started")
         elif game_to_join.max_players == len(game_to_join.players):
             raise Exception("Game is full")
-        # check if a player with the same name exists
-        elif game_to_join.players.exists(name=player.name):
+        # check if a player with the same name exists in the list
+        elif any(player.name == player_data.name for player in game_to_join.players):
             raise Exception("Player with same name exists")
 
-        player_created = game_to_join.players.create(player.model_dump())
-        player_created.table_position = len(game_to_join.players)
+        player = Player(**player_data.model_dump(), game=game_to_join)
+        player.table_position = len(game_to_join.players)
         # player_created contains the ponyorm object instance of the new player
         player.flush()  # flush the changes to the database
-        response = player_created.model_validate(player_created)
+        response = PlayerBase.model_validate(player)
     return response
 
 
@@ -36,7 +39,9 @@ def get_player(player_id: int, game_id: int):
     containing all the data from the player
     """
     with db_session:
-        player = games_crud.get_game(game_id).players[player_id]
+        player = Player.get(game=Game[game_id], id=player_id)
+        if player is None:
+            raise ObjectNotFound(Player, pkval=player_id)
         response = PlayerBase.model_validate(player)
     return response
 
@@ -47,7 +52,10 @@ def update_player(player: PlayerUpdate, game_id: int):
     and returns the PlayerBase schema with the updated data
     """
     with db_session:
-        player_to_update = games_crud.get_game(game_id).players[player.id]
+        game = Game[game_id]
+        player_to_update = Player.get(game=game, id=player.id)
+        if player_to_update is None:
+            raise ObjectNotFound(Player, pkval=player.id)
         player_to_update.set(**player.model_dump())
         player_to_update.flush()
         response = PlayerBase.model_validate(player_to_update)
@@ -60,6 +68,9 @@ def delete_player(player_id: int, game_id: int):
     and returns a message with the result
     """
     with db_session:
-        player = games_crud.get_game(game_id).players[player_id]
+        game = Game[game_id]
+        player = Player.get(game=game, id=player_id)
+        if player is None:
+            raise ObjectNotFound(Player, pkval=player_id)
         player.delete()
     return {"message": f"Player {player_id} deleted successfully"}
